# node_code/nodejs-server_gg/api/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Guided Gateway API",
    "description": "Move your app forward with the Uber API",
    "version": "1.0.0"
  },
  "produces": ["application/json"],
  "host": "localhost:8080",
  "basePath": "",
  "paths": {
  
  "/tours": {
  	
    "get": {
      "summary": "",
      "description":"Gets `Tour` objects.\nOptional query param of **size**, **city** , **theme** determines\nsize, location and theme of returned tour list\n",
      "x-swagger-router-controller": "Default",
      "tags": ["Default"],
      "operationId": "toursGet",
      "parameters": [
        {
  "name" : "size",
  "in" : "query",
  "description" : "Size of array",
  "required" : false,
  "type" : "number",
  "format" : "double"
},
        {
  "name" : "city",
  "in" : "query",
  "description" : "city of tour",
  "required" : false,
  "type" : "string"
},
        {
  "name" : "theme",
  "in" : "query",
  "description" : "themes of tour",
  "required" : false,
  "type" : "string"
}
        
      ],
      "responses": {
        "200": {
  "description" : "Successful response",
  "schema" : {
    "type" : "array",
    "items" : {
      "type" : "object",
      "properties" : {
        "single" : {
          "type" : "boolean"
        },
        "name" : {
          "type" : "string"
        }
      }
    }
  }
}
        
        
      }
    } 
    
  } ,
  
  "/place": {
  	
    "get": {
      "summary": "",
      "description":"Gets a specific `Place` object.\nOptional query param of **id** determines\nid of returned object\n",
      "x-swagger-router-controller": "Default",
      "tags": ["Default"],
      "operationId": "placeGet",
      "parameters": [
        {
  "name" : "tourid",
  "in" : "query",
  "description" : "ID of Place",
  "required" : false,
  "type" : "number",
  "format" : "double"
}
        
      ],
      "responses": {
        "200": {
  "description" : "Successful response",
  "schema" : {
    "type" : "object",
    "properties" : {
      "single" : {
        "type" : "boolean"
      },
      "name" : {
        "type" : "string"
      }
    }
  }
}
        
        
      }
    } 
    
  } ,
  
  "/places": {
  	
    "get": {
      "summary": "",
      "description":"Gets `place` objects.\nOptional query param of **size**, **theme** determines\nsize, location and theme of returned tour list\n",
      "x-swagger-router-controller": "Default",
      "tags": ["Default"],
      "operationId": "placesGet",
      "parameters": [
        {
  "name" : "size",
  "in" : "query",
  "description" : "Size of array",
  "required" : false,
  "type" : "number",
  "format" : "double"
},
        {
  "name" : "theme",
  "in" : "query",
  "description" : "themes of place",
  "required" : false,
  "type" : "string"
}
        
      ],
      "responses": {
        "200": {
  "description" : "Successful response",
  "schema" : {
    "type" : "array",
    "items" : {
      "type" : "object",
      "properties" : {
        "single" : {
          "type" : "boolean"
        },
        "name" : {
          "type" : "string"
        }
      }
    }
  }
}
        
        
      }
    } 
    
  } ,
  
  "/search": {
  	
    "get": {
      "summary": "",
      "description":"Gets `search results` objects.\nOptional query param of **size**, **keyword** determines\nsize, search keyword , search criteria\n",
      "x-swagger-router-controller": "Default",
      "tags": ["Default"],
      "operationId": "searchGet",
      "parameters": [
        {
  "name" : "size",
  "in" : "query",
  "description" : "Size of array",
  "required" : false,
  "type" : "number",
  "format" : "double"
},
        {
  "name" : "keyword",
  "in" : "query",
  "description" : "themes of keyword",
  "required" : false,
  "type" : "string"
},
        {
  "name" : "criteria",
  "in" : "query",
  "description" : "criteria of search",
  "required" : false,
  "type" : "string"
}
        
      ],
      "responses": {
        "200": {
  "description" : "Successful response",
  "schema" : {
    "type" : "array",
    "items" : {
      "type" : "object",
      "properties" : {
        "single" : {
          "type" : "boolean"
        },
        "name" : {
          "type" : "string"
        }
      }
    }
  }
}
        
        
      }
    } 
    
  } ,
  
  "/guide": {
  	
    "get": {
      "summary": "",
      "description":"Gets a specific `Guide` object.\nOptional query param of **id** determines\nindividual guide\n",
      "x-swagger-router-controller": "Default",
      "tags": ["Default"],
      "operationId": "guideGet",
      "parameters": [
        {
  "name" : "id",
  "in" : "query",
  "description" : "id of guide",
  "required" : false,
  "type" : "number",
  "format" : "double"
}
        
      ],
      "responses": {
        "200": {
  "description" : "Successful response",
  "schema" : {
    "type" : "object",
    "properties" : {
      "name" : {
        "type" : "string"
      },
      "id" : {
        "type" : "number"
      }
    }
  }
}
        
        
      }
    } 
    
  } ,
  
  "/tour": {
  	
    "get": {
      "summary": "",
      "description":"Gets a specific `Tour` object.\nOptional query param of **tourid** determines\nid of returned object\n",
      "x-swagger-router-controller": "Default",
      "tags": ["Default"],
      "operationId": "tourGet",
      "parameters": [
        {
  "name" : "tourid",
  "in" : "query",
  "description" : "ID of tour",
  "required" : false,
  "type" : "number",
  "format" : "double"
}
        
      ],
      "responses": {
        "200": {
  "description" : "Successful response",
  "schema" : {
    "type" : "object",
    "properties" : {
      "single" : {
        "type" : "boolean"
      },
      "name" : {
        "type" : "string"
      }
    }
  }
}
        
        
      }
    } 
    
  } ,
  
  "/guides": {
  	
    "get": {
      "summary": "",
      "description":"Gets `Guide` objects.\nOptional query param of **size** , **city**, **speciality** determines\nsize of returned guide list\n",
      "x-swagger-router-controller": "Default",
      "tags": ["Default"],
      "operationId": "guidesGet",
      "parameters": [
        {
  "name" : "size",
  "in" : "query",
  "description" : "Size of array",
  "required" : false,
  "type" : "number",
  "format" : "double"
},
        {
  "name" : "city",
  "in" : "query",
  "description" : "city of tour",
  "required" : false,
  "type" : "string"
},
        {
  "name" : "theme",
  "in" : "query",
  "description" : "theme of guide",
  "required" : false,
  "type" : "string"
}
        
      ],
      "responses": {
        "200": {
  "description" : "Successful response",
  "schema" : {
    "type" : "array",
    "items" : {
      "type" : "object",
      "properties" : {
        "name" : {
          "type" : "string"
        },
        "id" : {
          "type" : "number"
        }
      }
    }
  }
}
        
        
      }
    } 
    
  } 
  
   
  }, "definitions": {
  
  }
}